Update LUUtils.ps1 (#7)
* Update LUUtils.ps1

This will avoid the bug from occurring due to the recognizers having names as substrings of another recognizer. 
The "/" in front of the luModel name will help to avoid taking the recognizers which substring names such as "renewLicense" which has newLicense in it. 
In the current scenario, both the recognizers are returned which will throw an error while running the pipeline as the path could not be found as there will be two recognizer paths in the same line.

* Update LUUtils.ps1

* Update LUUtils.ps1